requirements.txt
Django~=5.0.6
aiogram~=3.6.0
python-dotenv~=1.0.1
asgiref~=3.8.1
requests~=2.31.0
pytz~=2024.1
psycopg2-binary
pip~=23.2.1
attrs~=23.2.0
idna~=3.7
multidict~=6.0.5
numpy~=1.26.4
sqlparse~=0.5.0
pandas~=2.2.2
openpyxl~=3.1.4
python-dateutil~=2.9.0.post0
pydantic~=2.7.1
aiofiles~=23.2.1
certifi~=2024.2.2
aiohttp~=3.9.5
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
urllib3~=2.2.1
six~=1.16.0
djangorestframework~=3.15.2